refactor: product 도메인에 종속된 createFetchProductsParams를 fetchProducts 내부로 이동
