+++ src/utils/index.ts
@@ -1,5 +1,4 @@
 export { default as createFetchUrl } from "./createFetchUrl";
-export { default as createFetchProductsParams } from "./createFetchProductsParams";
 export { default as generateBasicToken } from "./generateBasicToken";
 export { default as convertToLocaleAmount } from "./convertToLocaleAmount";
 export { default as validateProductImageUrl } from "./validateProductImageUrl";
